Added E button.

+++ src/input/button/button_state.rs
@@ -16,7 +16,7 @@ impl ButtonState {
             _ => false,
         }
     }
-    pub fn into_bits(self) -> u8 {
+    pub fn into_bits(self) -> u16 {
         match self {
             ButtonState::Released => 0b00,
             ButtonState::JustReleased => 0b10,
@@ -25,7 +25,7 @@ impl ButtonState {
         }
     }
 
-    pub fn from_bits(value: u8) -> Option<Self> {
+    pub fn from_bits(value: u16) -> Option<Self> {
         match value {
             0b00 => Some(ButtonState::Released),
             0b10 => Some(ButtonState::JustReleased),
